switched avg_scattering_frac from spline to interp
This is necessary because for large grains and small wavelengths the
scattering is almost 100% in forward direction. The spline returns large
negative angles (e.g. -700) and this leads to a segmentation fauld
(which is correct behaviour). There were no seg faults in earlier
versions because the getScatThetaID function just returned to maximum
value. The scattering was wrong with the old version. Most of the time
this affects only the temperature simulation but in principle could
affect all types of simulations with huge forward scattering.